ECLTransmissibility: Prepare for Cylinder Coordinates
This commit is a first step in preparing class EclTransmissibility
for a (simplified) calculation of transmissibilities using method
OLDTRAN on matching radial grids.  In particular, we refactor the
innards of member function update() to a new member function

    computeTrans_NewTran_()

and separate steps common to both the existing and the new methods
out to a new member function, prepareTranCalc_().  The latter
computes the net-to-gross ratio, defines the per-cell permeability
tensor and pre-allocates a reasonable amount of storage for the
final 'trans_' container.