ECLTransmissibility: Collect Element IDs in Structure
This commit adds a new private structure, EntityID, that holds the
compressed and the uncompressed IDs of a cell.  This does move the
calculation of uncompressed ID (i.e., cartMapper.cartesianIndex())
ahead of the inside.elemIdx > outside.elemIdx thereby making it
unconditional.  We may have to reconsider that choice if this
becomes a bottleneck.